+++ src/main/java/com/clickhouse/kafka/connect/sink/data/SchemaType.java
@@ -3,5 +3,6 @@ package com.clickhouse.kafka.connect.sink.data;
 public enum SchemaType {
     SCHEMA,
     SCHEMA_LESS,
-    STRING_SCHEMA
+    STRING_SCHEMA,
+    DEBEZIUM_CDC
 }